🚧 KSXT6H task: split runtime and backend hotspots
