+++ .agentplane/tasks/202607092208-KSXT6H/pr/meta.json
@@ -1,0 +1,13 @@
+{
+  "base": "main",
+  "branch": "task/202607092208-KSXT6H/split-runtime-and-backend-hotspots-for-v0-6-22",
+  "created_at": "2026-07-11T11:31:25.645Z",
+  "diffstat_sha256": "sha256:8660cc0ad4190fa3651eaa56e7bf1d354f9745871eeabfcef45717238e0fa3a8",
+  "last_verified_at": null,
+  "schema_version": 1,
+  "task_id": "202607092208-KSXT6H",
+  "updated_at": "2026-07-11T11:31:25.645Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
